lib/scheduler.py: fix the scheduler when obs are added on-the-fly
If the scheduler was running, and there were stills available, and the
`bin/add_observations_paper.py` script was used to ingest new observations
into the database, the scheduler would crash because it would try to locate
the action for observations with statuses of '' or 'NEW'. As best I can tell
the appropriate fix is to have get_action() return None for those statuses (as
it already does for 'COMPLETE' actions).